requirements.txt
python-telegram-bot[job-queue]
openai
requests
gtts
python-dotenv
